Minor tweaks

+++ src/p_local.h
@@ -89,7 +89,7 @@
 void P_SetupPsprites(player_t *player);
 void P_MovePsprites(player_t *player);
 void P_DropWeapon(player_t *player);
-void P_SetPsprite(player_t *player, int position, statenum_t stnum);
+void P_SetPsprite(player_t *player, size_t position, statenum_t stnum);
 
 //
 // P_USER
